Generacion claves OK

+++ funcionesNumeros.cpp
@@ -1,8 +1,8 @@
 #include "header.h"
 
-int getNumeroPrimo(int min, int max)                    //funcion que devuelve un numero primo entre determinado rango
+long int getNumeroPrimo(long int min, long int max)                    //funcion que devuelve un numero primo entre determinado rango
 {
-    int numeroPrimo;
+    long int numeroPrimo;
 
     do
     {
@@ -13,7 +13,7 @@ int getNumeroPrimo(int min, int max)                    //funcion que devuelve u
     return numeroPrimo;
 }
 
-int generarNumeroRandomRango(int min, int max)          //funcion que devuelve un numero random entre determinado rango
+long int generarNumeroRandomRango(long int min, long int max)          //funcion que devuelve un numero random entre determinado rango
 {
     int n = max - min + 1;
     int resto = RAND_MAX % n;
@@ -25,7 +25,7 @@ int generarNumeroRandomRango(int min, int max)          //funcion que devuelve u
     return min + x % n;
 }
 
-bool esPrimo(int numero)                                //funcion que devuelve 1 si numero es primo, sino devuelve 0
+bool esPrimo(long int numero)                                //funcion que devuelve 1 si numero es primo, sino devuelve 0
 {
     int i;
